Create 01_Searching_Programm.py
First you need to put text into a "text" variable and then by turning this code on it will ask what you searching for and if you, for example, type the letter "a" it will close all a's into () in the text.

Its some soft of "F3" function in Google Chrome Browser.